challenge_2/requirements.txt
fastapi
uvicorn
streamlit
streamlit-option-menu
langchain
langchain-community
langchain_google_genai
google-generativeai==0.8.1
python-dotenv